+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ### Unreleased
 
+### v0.28.0 - September 16, 2025
+
+- Support for experience syncing from a config file using `resim experiences sync ...`. In short, by
+  specifying which experiences they want with which tags, systems, and test suites in a yaml file, a
+  user can use this command to easily make the required updates so that the backend matches the
+  yaml.
+
+### v0.27.1 - September 12, 2025
+
+- Increase debug mode timeout to 30 minutes
+
 ### v0.27.0 - September 8, 2025
 
  - Extends Support for metrics sets to test suite creation and running